backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
openai
numpy
faiss-cpu
sentence-transformers
requests
dotenv